major: replace uint32 packet data with raw bytes, add on_connected callback to BLE backend

+++ include/ndx/acquisition_backend.hpp
@@ -9,7 +9,7 @@
 namespace ndx {
 
 struct Packet {
-  std::vector<uint32_t> data;
+  std::vector<uint8_t> data;
   std::uint64_t timestamp_ms;
 };
